src/requirements.txt
annotated-types==0.7.0
click==8.3.1
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
flask-openapi3==4.2.0
flask-openapi3-swagger==5.25.3
Flask-SQLAlchemy==2.5.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
six==1.17.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==2.3.8
